Fixing the tex formatting of 'and' in declarations
The TeX parser inserts the expository term 'and' before the last term in a declaration sequence, but this causes ambiguity when the conjunction operator 'and' is one of the constants being declared.  So we change the color of the expository 'and' to black even when viewing meaning, so only the meaningful terms are gold.