Showing global constant declarations in the context.
This is a quickie first pass addition that gathers together all of the constant declarations and puts them in a simple list at the top of the context.  Note that it currently doesn't work if there is no context.  In the future, this should be moved to a separate mechanism that gives detailed reports on the documents validity.