Making the Validation notification modal.
The validation notification being a TinyMCE notification has two problems.  First, it allows the user to edit the document while validation is going on, so that the resulting feedback icons might be incorrect.  Second, if the user has an expression editor dialog open when validation completes, TinyMCE closes both the notification and the dialog, which is annoying.