Eliminating application settings that are no longer needed.
There was a bug in the Applications Settings dialog that was caused by trying to eliminate deprecated settings that the user no longer needs to change.  This restores that functionality by effectively removing many of the settings that are no longer required.  There is still leftover code that depends on those deprecated settings and thus can no longer ever be executed and should be cleaned up at some point, but this fixes the bug for now.