Just One Weenie
Added the feature that allows validation to only instantiate one weenie per rule per pass, resulting in a 55% speed improvement in validation.  In order for this to work we must split multi-conclusion rules into one rule per conclusions.  We also have to check for doppelgangers of the chosen weenie (that differ from it only by a permutation of the metavariables) and instantiate that on the same pass as well.