Support optional formatting of tuples as column vectors via appended tick mark.
Updated parsers to allow an option tick mark after the closing bracket of a tuple to format it (tex) as a column vector.  The meaning (putdown) is unaffected. For a matrix the optional tick mark can only be applied to the outer bracket, and if present formats it as the transpose of the matrix is would be without the tick mark (i.e. a horizonalt sequence of column vectors).  The meaning (putdown) of the matrix is unaffected.